Internalize test suite
